+++ src/transformation.cpp
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: BSD-3-Clause
 
-#include <spatial_transformation/transform.hpp>
+#include <spatial_transformation/transformation.hpp>
 
 namespace spatial_transformation
 {
